remove collector config tool legacy functionality
This change removes functionality from the web based collector config tool in
preparation for moving that functionality elsewhere.

The custom collector pipeline that the previous version could instantiate, where
each component was wrapped in a websocket repeater, is being moved to a websocket
processor. The UI elements that could display information received from those
websocket connections is being moved to a small web application that will
run via an extension.

What remains here will effectively be a static web page to help users configure
(or reconfigure) a collector.

This is the first change of many, and it just provides a skeletal implementation
that doesn't actually do anything yet.