move otto to its own dir, create c3 dir
The Otto proof of concept is usable but will be superseded by the WebSocket Processor,
WebSocket Processor Viewer Extension, and what for now is tentatively called the Collector
Component Configurator (c3). This change moves the Otto module into its own directory,
and the beginnings of c3 code (in development) to the c3 directory.